requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.14
dlib==19.24.5
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.65.4
idna==3.7
numpy==1.24.4
opencv-python==4.10.0.84
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1-modules==0.4.0
requests==2.32.3
rsa==4.9
sqlparse==0.5.1
typing-extensions==4.12.2
urllib3==2.2.2
